requirements.txt
openai
fastapi
uvicorn
pymupdf
python-dotenv
tiktoken
